When reconnecting disassociate dc and apnContext.
This is done in onActionIntentReconnectAlarm, which is called when
retrying after an error, by setting the ApnContext.DataConnection &
DataConnctionAc to null which intern clears the DataConnection.mApnList.

With the mApnList cleared DataConnection.mRefCount will not be incremented
because the DC will not be found by findReadyDataConnection. Thus fixing
at least one cause of 6529189.

Bug: 6529189
Change-Id: I40ff1966f6fbd2f69e97d77b9db7122e4c5b992f
Preserve retryCount when retrying.

Bug: 6529189
Change-Id: I743aed08aa161d91899aeb2ccacffcd224e48c0a
Fix our interpretation of apn setting's bearer.

Between ICS and JB we changed it to use TelephonyManager's network type
enumeration.  This changes it back to represent the RIL's radio
technology.  The only diff between the two is 13 and 14 are swapped,
so the change was causing problems with LTE and eHRPD.

bug:6830908
Change-Id: I85a7c8037f370f368e5cbf6dc2c9395c6889b541
Fix detection of CDMA cell broadcast messages.

The bIsServicePresent field of an incoming CDMA SMS RIL message
is not a reliable way to distinguish broadcast from point-to-point
messages on some RILs. Change CDMA SmsMessage.getMessageType() to
return MESSAGE_TYPE_BROADCAST when the service category is not 0,
and MESSAGE_TYPE_POINT_TO_POINT when the service category is 0.

Bug: 6853691
Change-Id: I0d44a8ce0bb295da5125271db90652570d655bc1
Implement CMAS service category program results.

Correctly handle CDMA Service Category Program Data requests,
and send the SCPT response to the message center. Parcel SCPD
operations as an ArrayList instead of an array (current version
will throw ClassCastException when the array is cast).

Bug: 6853691
Change-Id: I949ea68891c78306059248832e59a593ab606e11
Fix 6804479: "Emergency calls only" in notification shade

Add an additional codepath to the "no connectivity" logic in
NetworkController to allow the PLMN bubble up from the
GsmServiceStateTracker, where R.string.emergency_calls_only
is returned if we're in emergency call mode.

Bug: 6804479
Change-Id: I0a77261e4393cc0dc32bae3e631ef196b2342f06
Fix NPE on some phones at startup.

Bug: 6877589
Change-Id: I65ff37b4cf6b59a72a7aa665f252f5fb34f3b183